requirements.txt
Flask
simplejson
psycopg2-binary